+++ frontend/src/components/index.ts
@@ -3,7 +3,11 @@ export * from 'components/@shared/FlexBox';
 export * from 'components/@shared/Button';
 export * from 'components/@shared/ModalOverlay';
 export * from 'components/@shared/DarkModeButton';
-export * from 'components/Input';
+export * from 'components/@shared/FixedButton';
+export * from 'components/@shared/InfiniteScroll';
+export * from 'components/@shared/BottomSheet';
+export * from 'components/@shared/Dropdown';
+export * from 'components/@shared/Input';
 export * from 'components/LinkText';
 export * from 'components/Logo';
 export * from 'components/ValidationMessage';
@@ -16,7 +20,6 @@ export * from 'components/CheckCircles';
 export * from 'components/Timer';
 export * from 'components/UnderLineText';
 export * from 'components/ChallengeIcon';
-export * from 'components/@shared/FixedButton';
 export * from 'components/Profile';
 export * from 'components/EmptyContent';
 export * from 'components/SnackBar';
@@ -26,14 +29,11 @@ export * from 'components/LandingNavigation';
 export * from 'components/ToggleButton';
 export * from 'components/Title';
 export * from 'components/LoadingButton';
-export * from 'components/@shared/InfiniteScroll';
 export * from 'components/Challenger';
 export * from 'components/ChallengerList';
 export * from 'components/ChallengeDetail';
 export * from 'components/CommentItem';
 export * from 'components/FeedItem';
-export * from 'components/@shared/BottomSheet';
-export * from 'components/@shared/Dropdown';
 export * from 'components/Bell';
 export * from 'components/SubscriptionButton';
 export * from 'components/CustomCycleTimeBottomSheet';
